Finished with authentication and authorization

+++ frontend/src/components/Workout.jsx
@@ -1,13 +1,20 @@
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const Workout = ({ workout }) => {
   const { dispatch } = useWorkoutsContext()
+  const { user } = useAuthContext()
 
   const handleClick = async () => {
+    if(!user) return
+    
     const { _id: id } = workout
     const response = await fetch(`http://localhost:5000/api/v1/workouts/${ id }`, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     })
     const jsonData = await response.json()
